move test out of test_viral_snippets
